backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
python-dotenv==1.0.0
yt-dlp==2023.10.13
openai==1.3.5
google-cloud-texttospeech==2.14.1
pytest==7.4.2
celery==5.3.4
redis==5.0.1
python-json-logger==2.0.7
requests==2.31.0
pydantic==2.4.2